Make memDB the default DB option, and hide SqlLite behind a build tag.
